UHF-9961: Remove preview controller, add preview drush command
